// src/modules/mars/interfaces/ICardActivateResponse.interface.ts
export interface ICardActivateResponse {
  attractionChips: number;
  gameChips: number;
  pointsToNextReward: number;
  rewardChips: number;
  rewardPoints: number;
  tickets: number;
  videoChips: number;
  storeId: number;
  cardNumber: string;
  country: string;
  cardEncoding: string;
  cardStatusId: number;
  isRegistered: boolean;
}